Changed some stuff (u32 -> u16) and dealt with mutable references

+++ src/components/instruction_set.rs
@@ -21,7 +21,7 @@ pub enum Opcodes{
     OP_TRAP    /* execute trap */
 }
 
-impl From<Opcodes> for i32 {
+impl From<Opcodes> for u16 {
     fn from(value: Opcodes) -> Self {
         match value{
             Opcodes::OP_BR=> 0,
